requirements.txt
google-genai
python-dotenv
pydantic
discord-webhook
discord.py